+++ src/main/java/necessecompanions/mobs/PenguinPetMob.java
@@ -1,0 +1,49 @@
+package necessecompanions.mobs;
+
+import necesse.engine.gameLoop.tickManager.TickManager;
+import necesse.engine.registries.MobRegistry;
+import necesse.entity.mobs.MaskShaderOptions;
+import necesse.entity.mobs.MobDrawable;
+import necesse.entity.mobs.PlayerMob;
+import necesse.gfx.camera.GameCamera;
+import necesse.gfx.drawOptions.DrawOptions;
+import necesse.gfx.drawables.OrderableDrawables;
+import necesse.gfx.gameTexture.GameTexture;
+import necesse.level.maps.Level;
+import necesse.level.maps.light.GameLight;
+import necessecompanions.ModEntryPoint;
+import necessecompanions.mobs.base.AbstractCompanionMob;
+
+import java.awt.*;
+import java.util.List;
+
+public class PenguinPetMob extends AbstractCompanionMob {
+    public static GameTexture texture;
+    public static boolean hasMultipleSprites;
+
+    public PenguinPetMob() {
+        super(100, ModEntryPoint.SWIM_SPEED_BUFF);
+    }
+
+    @Override
+    public void addDrawables(List<MobDrawable> list, OrderableDrawables tileList, OrderableDrawables topList, Level level, int x, int y, TickManager tickManager, GameCamera camera, PlayerMob perspective) {
+
+        GameLight light = level.getLightLevel(x / 32, y / 32);
+        int drawX = camera.getDrawX(x) - 16;
+        int drawY = camera.getDrawY(y) - 28;
+        int dir = this.getDir();
+        Point sprite = this.getAnimSprite(x, y, dir);
+        drawY += this.getBobbing(x, y);
+        drawY += this.getLevel().getTile(x / 32, y / 32).getMobSinkingAmount(this);
+        final MaskShaderOptions swimMask = this.getSwimMaskShaderOptions(this.inLiquidFloat(x, y));
+        final DrawOptions options = MobRegistry.Textures.penguin.initDraw().sprite(sprite.x, sprite.y, 32).addMaskShader(swimMask).light(light).pos(drawX, drawY);
+        list.add(new MobDrawable() {
+            public void draw(TickManager tickManager) {
+                swimMask.use();
+                options.draw();
+                swimMask.stop();
+            }
+        });
+        this.addShadowDrawables(tileList, x, y, light, camera);
+    }
+}